[Core reorg] Organize source files per area.
This is based on pull request #431 with minor changes.
