+++ README.md
@@ -6,9 +6,13 @@ Developers mailing list: https://dev.eclipse.org/mailman/listinfo/wakaama-dev
 ## Source Layout
 
     -+- core                   (the LWM2M engine)
+     |
+     +- coap                   (CoAP stack adaptation)
      |    |
      |    +- er-coap-13        (Modified Erbium's CoAP engine from
      |                          https://web.archive.org/web/20180316172739/http://people.inf.ethz.ch/mkovatsc/erbium.php)
+     |
+     +- data                   (data formats serialization/deserialization)
      |
      +- tests                  (test cases)
      |
