+++ coap/er-coap-13/er-coap-13.c
@@ -47,7 +47,7 @@
 
 #include "er-coap-13.h"
 
-#include "../liblwm2m.h" /* for lwm2m_malloc() and lwm2m_free() */
+#include "liblwm2m.h" /* for lwm2m_malloc() and lwm2m_free() */
 
 #define DEBUG 0
 #if DEBUG
