+++ examples/lightclient/CMakeLists.txt
@@ -7,6 +7,8 @@ if(DTLS)
 endif()
 
 include(${CMAKE_CURRENT_LIST_DIR}/../../core/wakaama.cmake)
+include(${CMAKE_CURRENT_LIST_DIR}/../../coap/coap.cmake)
+include(${CMAKE_CURRENT_LIST_DIR}/../../data/data.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/../shared/shared.cmake)
 
 add_definitions(-DLWM2M_CLIENT_MODE)
@@ -16,7 +18,7 @@ endif()
 
 add_definitions(${SHARED_DEFINITIONS} ${WAKAAMA_DEFINITIONS})
 
-include_directories (${WAKAAMA_SOURCES_DIR} ${SHARED_INCLUDE_DIRS})
+include_directories (${WAKAAMA_HEADERS_DIR} ${COAP_HEADERS_DIR} ${DATA_HEADERS_DIR} ${WAKAAMA_SOURCES_DIR} ${SHARED_INCLUDE_DIRS})
 
 SET(SOURCES
     ${CMAKE_CURRENT_LIST_DIR}/lightclient.c
@@ -26,7 +28,7 @@ SET(SOURCES
     ${CMAKE_CURRENT_LIST_DIR}/object_test.c
     )
 
-add_executable(${PROJECT_NAME} ${SOURCES} ${WAKAAMA_SOURCES} ${SHARED_SOURCES})
+add_executable(${PROJECT_NAME} ${SOURCES} ${WAKAAMA_SOURCES} ${COAP_SOURCES} ${DATA_SOURCES} ${SHARED_SOURCES})
 
 # Add WITH_LOGS to debug variant
 set_property(TARGET ${PROJECT_NAME} APPEND PROPERTY COMPILE_DEFINITIONS $<$<CONFIG:Debug>:WITH_LOGS>)
